refactor: split files apart
